[update-checkout][master-next] Update repos that actually have master-next branches to check these out.
The reason that master-next branches for these repositories exist is that
swift's master-next branch is slightly behind swift's master branch.

If swift's master branch causes changes to be made to this branch due to an
upstream change, there is no guarantee that these changed repos will compile
against master-next. In effect this allows to have a delayed snapshot of the
master branch of these repos.