test: correct DebugInfo test
Fix a test case which did not test what it thought it did.  The `-SAME-NOT`
suffix is not supported.  Make the test more strict to check the entire entry to
ensure that errant flags are not propagated instead.  This makes the test less
resilient to encoding changes, but ensures that the test performs the desired
check.  NFC.