basic: introduce the _environment compilation conditional
Windows has multiple environment flavours, each of which are different.  cygwin
provides a glibc based environment, MinGW provides a msvcrt based GNU
environment, itanium provides a msvc environment with an itanium C++ ABI, and
msvc is the pure msvc, msvcprt based environment as vended by Microsoft.  Due to
the differences in the environment, we need a means to differentiate them at the
stdlib level.  Provide the _environment conditional which would allow you to
query what environment is being targeted.