[cmake] Change all of the *_{BUILD,INCLUDE}_* options into their own section. NFC.
In order to reduce build times for LTO, I am going to "harmonize" these options
with LLVM. This is just a cleanup commit.

When I say "harmonize" I mean that currently, we haphazardly use either
SWIFT_BUILD_* or SWIFT_INCLUDE_* to not include/build targets. This I believe is
due to cargo-culting from LLVM without understanding the *REAL MEANING* of these
variables.

The *REAL MEANING* of these sorts of variables are:

1. SWIFT_BUILD_${X} means that cmake should generate build targets for ${X} and
build those tools by default.
2. SWIFT_INCLUDE_${X} means that cmake should only generate build targets for ${X}.