stdlib: shim errno access
errno is implemented as a macro in many environments.  The accessor hidden
behind the macro is not a standard function, so we ended up with an
implementation specific handling across all the targets.  Shim the function in C
where it can be hidden behind the CPP.  This simplifies the implementation on
the swift side.