[sil-printer] Only sort the PredIds, UserIDs that we put in comments if we are supposed to emit sorted SIL.
This was done some time ago to make it easier to diff large amounts of SIL
output. The problem is that it makes it difficult to know the *true* memory
order that the predecessor list is in which can lead to surprise when working
with SIL and create test cases.

I believe some time after that point we added the notion of "sorted" sil, i.e.
SIL that does not guarantee any relation to the actual memory representation of
the SIL and is meant to ease diffing. This fits nicely with the true intention
of this sort of sorting.

Thus this commit puts sorting PredIDs, UserIDs behind that flag.