[lit] Rename config.targets_to_build => config.llvm_code_generators and allow it to actually take on a value.
Previously this variable was always just "". Now using the recent cmake changes
that I made, we can just grab the proper values from LLVMConfig.cmake.

I need this so that we can REQUIRES tests based on the code generators that
LLVM was compiled with.

To make it easy to show the dependency in the requires line, the requires
pattern to use is:

REQUIRES: CODEGENERATOR=...