[build-script] Change LLVM_TARGETS_TO_BUILD to be a configurable option from the command line.
While it would be great to leave this as an option only configurable by a change
to the script, it is important that we make this configurable from the command
line so that we can reduce the work needed to be done when compiling with full
LTO.

This is trivially correct due to the way we convert build-script-impl args to
variables (i.e. llvm-targets-to-build => LLVM_TARGETS_TO_BUILD).

rdar://24717107