[build-script] When smoke testing on macOS, only build the X86 LLVM code generator
We have no need for the other LLVM code generators (after all the main principle
of this smoke test is to only text X86 on macOS).

To implement this I had to split a couple of IRGen tests that involved multiple
code generator into multiple tests so I could use different REQUIRES lines. This
would be a nice feature to add to lit.