[cmake] Do not build/link parts of LLVM that we do not use if we are not going to test LLVM.
These are parts of LLVM that Swift strictly /never/ uses, but are used by LLVM
for testing. In a future commit, functionality will be added to build-script to
test LLVM.

Other parts of LLVM that can be used by Swift depending on ones configuration
will be disabled in a future patch via the introduction of a new components
system.