Speculative fix for code-completion nullptr dereference
I haven't managed to find a test case, but we see crash reports about
this call. It may be that the correct fix is for some other part of the
compiler not to produce a null protocol self, but it's harmless for the
ASTPrinter to dodge a null pointer here.

rdar://problem/27053440