Make the default access level 'internal'.
Now that we're equipped to check the access level properly, taking
parent scopes into account, flip the switch as specified by SE-0025:
the default access level everywhere is 'internal'.

The one exception is constrained extensions, as described in 306eddab.
I haven't thought through whether these are safe or not, so for now
I'm leaving the logic conservative. If it turns out to be safe I'll
consider it a bug fix.

Part of SE-0025 ('fileprivate' and 'private').