SILOptimizer: Fix a broken test
The test only passed on accident with -O prior to the
'@noreturn -> Never' change, because '@noreturn' in a
function_ref matched the 'CHECK: return' line.

Make the CHECK: more specific, which even before my
changes exposes that there was a strong_retain in
there in optimized builds.

Fixes <rdar://problem/27506150>.